[DEV-3443] Minor edits when reviewing PR
